Added Game object property to Board class to hold the game the board is in. Added Game parameter to Board constructor.
Made some changes to the Game class. Renamed the old playerTurn() function that runs in the command line to playerTurnCmd(). Added new playerTurn() function only takes an integer array for coordinates.

Added event handlers to each BoardSpace object created in initBoard() function in Board class. Event handlers call the playerTurn() function of the board's game object. Passes the boardSpace's GridSpace as the coord parameters.

Added currentPlayer property to the Game class to hold the current player. Added the switchPlayer() function to the Game class to switch the players. This function is called at the end of player turn to switch players.